+++ aiograpi/mixins/public.py
@@ -25,6 +25,7 @@ from aiograpi.exceptions import (
     TermsAccept,
     TermsUnblock,
 )
+from aiograpi.utils.logging import truncate_log_text
 from aiograpi.utils.timing import random_delay
 
 
@@ -224,7 +225,7 @@ class PublicRequestMixin:
                 "Status %s: JSONDecodeError in public_request (url=%s) >>> %s",
                 response.status_code,
                 url,
-                response.text,
+                truncate_log_text(response.text),
             )
             raise ClientJSONDecodeError(
                 "JSONDecodeError {0!s} while opening {1!s}".format(e, url),